% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year.R
\name{add_status_year}
\alias{add_status_year}
\title{Add status year column to collection}
\usage{
add_status_year(collection, item_status_date_column = "ItemStatusDate")
}
\arguments{
\item{collection}{A data frame containing a collection.}

\item{item_status_date_column}{The name of the column containing the item status date.}
}
\value{
The \code{collection} data frame with \code{status_year} adjoined.
}
\description{
A function to adjoin the year from the item status date to the collection database.
}
\examples{
taxon_names = c('Trigonella afghanica', 'Eupatorium magdalenae')
item_status = c('2021-04-12', '12th August 1991')
collection = data.frame(name = taxon_names, item_status = item_status)
add_status_year(collection, item_status_date_column ='item_status')
}
