% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_information.R
\name{add_wcvp_distributions}
\alias{add_wcvp_distributions}
\title{Add wcvp_distributions information to wcvp_names}
\usage{
add_wcvp_distributions(filepath = NULL, wcvp, use_rWCVPdata = FALSE)
}
\arguments{
\item{filepath}{path to wcvp_distributions.csv file.}

\item{wcvp}{Output from import_wcvp_names}

\item{use_rWCVPdata}{Flag for whether we use rWCVPdata to get wcvp_distributions}
}
\value{
Desired information from wcvp_distributions.csv combined with wcvp_names.
}
\description{
We allow three methods for uploading the information from wcvp_distributions:
\itemize{
\item By choosing a file path to a csv file containing the information using \code{filepath}.
\item By loading an R save (file type \code{.RData}, \code{.rdata}, \code{.rda} or \code{.rds}) via \code{filepath}.
\item By using the package  \code{rWCVPData}, by setting \code{use_rWCVPData} = TRUE.
}
}
