% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{author_words}
\alias{author_words}
\title{Extract words from author}
\usage{
author_words(author)
}
\arguments{
\item{author}{Author that wants splitting into words.}
}
\value{
Words contained in the author.
}
\description{
Extract words from author
}
\details{
From the string \code{author} we extract any words that begin with a capital and is followed by at least two lower case letters or hyphens. Or the special cases \code{DC.}, \code{Sm.} and \code{Br.}. In the special cases \verb{\\\\} will be added before \code{.} to allow regular expression searching of the authors such as \code{grepl("DC\\\\.", "Some Author")}. We have not included \code{L.} since this could be mistaken for an initial of another author rather than Carl Linnaeus.

This function is used when comparing authors in \code{\link[=author_check]{author_check()}}.
}
\examples{
author_words('(Jacq.) Schott')
author_words('Villarroel & J.R.I.Wood')
author_words('(DC.) F.Muell.')
}
