% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_to_wcvp.R
\name{convert_to_accepted_name}
\alias{convert_to_accepted_name}
\title{Convert to accepted plant name for World Checklist of Vascular Plants}
\usage{
convert_to_accepted_name(original_match, wcvp)
}
\arguments{
\item{original_match}{The indices of the rows in \code{wcvp$wcvp_names} (or generally referred to as enrich_database) corresponding to the matches of the collection to it.}

\item{wcvp}{World Checklist of Vascular Plants (WCVP) database, obtained using the function \code{\link[=import_wcvp_names]{import_wcvp_names()}}.}
}
\value{
match and message after converting to accepted name
}
\description{
Convert to accepted plant name for World Checklist of Vascular Plants
}
