% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Diversity_classification}
\alias{Diversity_classification}
\title{Diversity_classification}
\format{
An object of class \code{list} of length 2.
}
\usage{
Diversity_classification
}
\description{
A list containing two dataframes:
\itemize{
\item Classification: containing how families link to higher order groups.
\item synomym_families: containing family names that are synomym and which to use within Classificiation.
}
}
\keyword{datasets}
