% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules-utilities.R
\name{exist_at_date}
\alias{exist_at_date}
\title{Generate data frame of records that exist each year}
\usage{
exist_at_date(
  date,
  AccessionYear,
  ItemStatusDate,
  ItemStatusType,
  post_date = as.character(Sys.Date())
)
}
\arguments{
\item{date}{dates in the format \code{YYYY-MM-DD} or \code{YYYY/MM/DD}.}

\item{AccessionYear}{The accession year of the records.}

\item{ItemStatusDate}{The data when the records were last updated.}

\item{ItemStatusType}{The status of the record when last updated. Either \code{Existing} or \code{NotExisting}.}

\item{post_date}{Set the date that existing plants in the LC are classed as alive until . Default is the present day.}
}
\value{
data frame where each column corresponds to a date, each row an item in the collection. The (i,j)th value is a logical (TRUE/FALSE) corresponding whether the ith item existed at the jth date.
}
\description{
This function calculates which items in a collection are existing at a given date (or dates).
}
\details{
Accession date is inputted through \code{AccessionYear} which accepts only the year, for example \code{c(2014,2017,2015)}. The accession date is then set to the 1st of January in the given year.

Item death is calculated via the inputs \code{ItemStatusDate} and \code{ItemStatusType}. If the item is still alive (\verb{ItemStatusType == 'Existing}) then the death date is set to today (\code{Sys.Date()}). Otherwise the item is dead (\verb{ItemStatusType == 'NotExisting}) and we use the corresponding date given in \code{ItemStatusDate}. If only the year is available in \code{ItemStatusDate} we set the date of death to the 31st of December that year. If only the month and year is available we set the day of the item death to be the 28th.

\code{AccessionYear}, \code{ItemStatusDate} and \code{ItemStatusType} inputs must have the same length and the ith value in each corresponds to the ith item in the collection.
}
\examples{
exist_at_date(date = c('2003-06-25', '2010-03-07'),
AccessionYear = c('1956', '1988', '2005', '2018'),
ItemStatusDate = c('2004-01', '2010-03-05', '2022-04-08', '2022-08-19'),
ItemStatusType = c('NotExisting','NotExisting','Existing','Existing'))
}
