% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_year.R
\name{extract_year}
\alias{extract_year}
\title{Extract the year from different date formats}
\usage{
extract_year(dates)
}
\arguments{
\item{dates}{A vector of dates.}
}
\value{
A vector of corresponding years.
}
\description{
A function that extracts the year from dates where the date can be in such formats as: \code{DD/MM/YYYY}, \code{YYYY-MM-DD}, \code{MM/YYYY}, \code{YYYY-MM}, \code{YYYY}, \verb{DD Month YYYY}.
}
\details{
The method works by extracting any patterns of 4 numbers in a row.
Thus, the year must be in the format YYYY. I.e. cannot have formats such as DD/MM/YY.
}
\examples{
extract_year(c('03/05/2021', '3rd of May 2022', '2023-05-03'))
}
