% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_information.R
\name{generate_labels}
\alias{generate_labels}
\title{generate_labels()}
\usage{
generate_labels(data)
}
\arguments{
\item{data}{distribution data for plants taken from wcvp_distribution.}
}
\value{
labels
}
\description{
Generate shortened distribution details that can be used on labels in collections.
}
