% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_information.R
\name{get_accepted_plant}
\alias{get_accepted_plant}
\title{Get accepted plant details from POWO online}
\usage{
get_accepted_plant(powo_id)
}
\arguments{
\item{powo_id}{Plants of the World Online identifier.}
}
\value{
A vector of length two containing the accepted name and accepted powo_id.
}
\description{
This function queries POWO online for the accepted name of a plant and the corresponding Plants of the World Online identifier.
}
\details{
If the connection to POWO doesn't work (or exist) then the function returns (NA, NA).
}
\examples{
get_accepted_plant('582307-1')

}
