% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_information.R
\name{import_wcvp_names}
\alias{import_wcvp_names}
\title{Information from wcvp_names}
\usage{
import_wcvp_names(
  filepath = NULL,
  use_rWCVPdata = FALSE,
  wanted_columns = c("plant_name_id", "taxon_rank", "taxon_status", "family", "genus",
    "species", "lifeform_description", "climate_description", "taxon_name",
    "taxon_authors", "accepted_plant_name_id", "powo_id")
)
}
\arguments{
\item{filepath}{path to wcvp_names.csv file.}

\item{use_rWCVPdata}{Flag for whether we use rWCVPdata to get wcvp_names}

\item{wanted_columns}{specify extra columns to extract from \code{wcvp_names.csv}.}
}
\value{
Data frame of desired information from wcvp_names
}
\description{
We allow four three for uploading the information from wcvp_names:
\itemize{
\item By choosing a file path to a csv file containing the information using \code{filepath}.
\item By loading an R save (file type \code{.RData}, \code{.rdata}, \code{.rda} or \code{.rds}) via \code{filepath}.
\item By using the package  \code{rWCVPData}, by setting \code{use_rWCVPData} = TRUE.
}
}
\details{
The columns extracted always include:"plant_name_id", "taxon_name", "taxon_authors", "taxon_rank", "accepted_plant_name_id" and "powo_id". Further columns can be extracted using \code{wanted_columns} parameter, where either columns names or indices can be used.

We add two new columns:
\itemize{
\item 'taxon_length' which contains the number of characters in a given taxon_name used when searching for typos in garden databases.
\item 'is_autonym' contains a logical flag for whether the taxon name is an autonym. This is used to match autonyms in garden databases.
}

Moreover, the code performs fixes to the data provided namely:
\itemize{
\item Checks for '\link{*}' and '\link{**}' in taxon_name and taxon_rank, which should be 'var.' and 'f.' instead.
\item There are potentially plants that are synonyms without an accepted form in wcvp_names. However, in some cases the accepted form can be found on POWO online. Therefore, in this case we scrap POWO online for an accepted form and add to the loaded information in wcvp_names.
\item No author is provided for autonyms, so we extract the corresponding author from matching GenusSpecies record.
}
}
