% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LivingCollectionDynamics-package.R
\docType{package}
\name{LivingCollectionDynamics-package}
\alias{LivingCollectionDynamics}
\alias{LivingCollectionDynamics-package}
\title{LivingCollectionDynamics: Analyse Botanic Garden collections}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Jake Powell \email{jjp68@cam.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
