% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matching_functions.R
\name{match_authors}
\alias{match_authors}
\title{Find best author matches}
\usage{
match_authors(
  collection_author,
  enriched_database_authors,
  partial_method = "most words",
  ...
)
}
\arguments{
\item{collection_author}{The author from the collection wanted to be matched}

\item{enriched_database_authors}{Author options from the enrich_database.}

\item{partial_method}{Either \verb{'most words} or \code{'any words'}, defining the method used to find partial matches.}

\item{...}{Arguments (i.e., attributes) used in the matching algorithm (passed along to nested fuctions).}
}
\value{
a list of length 2 with:
\itemize{
\item \verb{$wanted} is a logical (TRUE/FALSE) vector with length \code{length(enriched_database_authors)} corresponding to the enriched database authors that most match the collection author.
\item \verb{$message} detailing whether the author match was exact, partial or no match was found.
}
}
\description{
Find best author matches
}
\examples{
collection_author = 'Schult'
enriched_database_authors = c("(Lour.) Schult", "Borhidi & E.Martinez")
match_authors(collection_author, enriched_database_authors)
}
