% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_enrich.R
\name{match_collection_to_enrich_database}
\alias{match_collection_to_enrich_database}
\title{Match collection to an enrichment database via taxonomic names}
\usage{
match_collection_to_enrich_database(
  collection,
  enrich_database,
  taxon_name_column = NA,
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  enrich_taxon_name_column = NA,
  enrich_taxon_authors_column = NA,
  typo_method = "All",
  do_add_split = TRUE,
  do_fix_hybrid = TRUE,
  do_rm_autonym = TRUE,
  do_rm_cultivar_indeterminates = TRUE,
  do_match_single = TRUE,
  do_match_multiple = TRUE,
  do_fix_taxon_name = TRUE,
  matching_criterion = LivingCollectionDynamics::no_additional_matching,
  ...
)
}
\arguments{
\item{collection}{A data frame containing a collection.}

\item{enrich_database}{A data frame of enriching information.}

\item{taxon_name_column}{The name of the column in the \code{collection} corresponding to taxonomic names.}

\item{taxon_name_full_column}{The name of the column in the \code{collection} corresponding to joined taxonomic names and authors.}

\item{taxon_author_column}{The name of the column in the \code{collection} corresponding to the authors of the taxonomic names.}

\item{enrich_taxon_name_column}{The name of the column in \code{enrich_database} that corresponds to the taxonomic names.}

\item{enrich_taxon_authors_column}{The name of the column in \code{enrich_database} that corresponds to the authors of taxonomic names.}

\item{typo_method}{Either \code{'All'}, \code{'Data frame only'},\code{'Data frame + common'}, \code{no}; detailing the level of typo finding required.}

\item{do_add_split}{Flag (TRUE/FALSE) for whether we search for missing f./var./subsp.}

\item{do_fix_hybrid}{Flag (TRUE/FALSE) for whether we search for hybrid issues.}

\item{do_rm_autonym}{Flag (TRUE/FALSE) for whether we try removing autonyms.}

\item{do_rm_cultivar_indeterminates}{Flag (TRUE/FALSE) for whether we remove cultivars and indeterminates prior to taxonomic name matching.}

\item{do_match_single}{Flag (TRUE/FALSE) for whether we do matching to unique taxonomic names in \code{enrich_database}.}

\item{do_match_multiple}{Flag (TRUE/FALSE) for whether we do matching to non-unique taxonomic names in \code{enrich_database}.}

\item{do_fix_taxon_name}{Flag (TRUE/FALSE) for whether attempt to fix common issues in taxonomic names to aid matching. Sections of common issue fixes can also be turned on/off using the inputs \code{do_add_split}, \code{do_fix_hybrid}, \code{do_rm_autonym}.}

\item{matching_criterion}{A function used to chose the best method from extracts of the \code{enrich_database}.}

\item{...}{Arguments (i.e., attributes) used in the matching algorithm (passed along to nested fuctions). Examples include, \code{enrich_display_in_message_column} and \code{enrich_plant_identifier_column}.}
}
\value{
A list of length seven containing:
\itemize{
\item \verb{$match} the index of the record in \code{enrich_database} which matches the record in the collection database.
\item \verb{$details_short} a simplified message detailing the match.
\item \verb{$match_taxon_name} a longer format message detailing the match.
\item \verb{$original_authors} The author/s (extracted) from the \code{collection} database.
\item \verb{$match_authors} The author/s of the matched record in \code{enrich_database}.
\item \verb{$author_check} Either \code{Identical}, \code{Partial} or \code{Different}  (\verb{No Match} if a match to \code{enrich_database} cannot be found). A message informing the similarity of the collection's taxon authors and the authors found in \code{enrich_database}. Author similarity is found using the function  \code{\link[=author_check]{author_check()}}.
}
}
\description{
Match collection to an enrichment database via taxonomic names
}
\details{
This function allows matching of a collection's database to an enrichment database.

By default the function uses all the steps of our matching algorithm, for details of this see the vignette \code{Matching.Rmd} (Method of Matching taxonomic records). If parts of the algorithm are not required these can be switched off using  \code{typo_method}, \code{do_add_split}, \code{do_fix_hybrid}, \code{do_rm_autonym}, \code{do_rm_cultivar_indeterminates}, \code{do_match_single}, \code{do_match_multiple} and \code{do_fix_taxon_name.} Moreover, by default no custom matching if performed. A user inputted custom matching criterion (function) can be added via the input \code{matching_criterion}.

To perform the matching you must specify the columns name of the taxon name in the enrichment database (\code{enrich_taxon_name_column}). If author matching is required then this column must also be specified for the enrichment database (\code{enrich_taxon_authors_column}).

The enrichment database must have some columns required for matching (\code{single_entry}, \code{taxon_length}, etc), we advice using \code{\link[=prepare_enrich_database]{prepare_enrich_database()}} to add these columns.

Similarly, you must specify the columns name of the taxon name in the collection database (\code{taxon_name_column}). If author matching is desired then you have two choices:
\itemize{
\item specify the taxon author column \code{taxon_author_column}.
\item Specify the combined taxon name and author column, \code{taxon_name_full_column} which removes words found in the taxon names from taxon names full to extract the authors.
}

Note if both are specified then the authors from \code{taxon_author_column} are used.
}
