% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_BGCI.R
\name{match_original_to_BGCI}
\alias{match_original_to_BGCI}
\title{match_original_to_BGCI()}
\usage{
match_original_to_BGCI(original_report, BGCI, taxon_name_col = "TaxonName")
}
\arguments{
\item{original_report}{original_report}

\item{BGCI}{BGCI}

\item{taxon_name_col}{taxon_name_col}
}
\value{
no gardens
}
\description{
match_original_to_BGCI()
}
