% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_enrich.R
\name{prepare_enrich_database}
\alias{prepare_enrich_database}
\title{Prepare enrichment database for matching}
\usage{
prepare_enrich_database(
  enrich_database,
  enrich_taxon_name_column = "taxon_names",
  enrich_taxon_authors_column = NA,
  enrich_taxon_name_full_column = NA,
  do_sanitise = TRUE,
  do_taxon_length = TRUE,
  do_single_entry = TRUE,
  do_author_parts = TRUE,
  do_add_id = TRUE,
  do_sort = TRUE,
  console_message = FALSE
)
}
\arguments{
\item{enrich_database}{A data frame of enriching information.}

\item{enrich_taxon_name_column}{The name of the column in \code{enrich_database} that corresponds to taxonomic names. Default value is \code{taxon_names}.}

\item{enrich_taxon_authors_column}{The name of the column in \code{enrich_database} that corresponds to the authors of taxonomic names.}

\item{enrich_taxon_name_full_column}{The name of the column in the \code{enrich_database} corresponding to joined taxonomic names and authors.}

\item{do_sanitise}{Flag (TRUE/FALSE) detailing whether to add the columns \code{sanitise_name}, \code{sanitise_author} and  \code{require_sanitise} corresponding to the sanitised taxonomic name, sanitised author and a flag (TRUE/FALSE) of whether the taxonomic name needed sanitising.}

\item{do_taxon_length}{Flag (TRUE/FALSE) detailing whether to add the column \code{taxon_length} containing the string length of the taxonomic names.}

\item{do_single_entry}{Flag (TRUE/FALSE) detailing whether to add the column \code{single_entry} containing whether the taxonomic name appears multiple times in \code{enrich_database}.}

\item{do_author_parts}{Flag (TRUE/FALSE) detailing whether to add the column \code{author_parts} containing the words extracted from the taxonomic author. This is used for partial author matching.}

\item{do_add_id}{Flag (TRUE/FALSE) detailing whether to add the column \code{ID} containing a unique identifer for each record in the \code{enrich_database}.}

\item{do_sort}{Flag (TRUE/FALSE) detailing whether to  alphabetically sort the taxonomic names in  \verb{enrich database}.}

\item{console_message}{Flag (TRUE/FALSE) detailing whether to show messages in the console.}
}
\value{
\code{enrich_database} with additional columns used by matching to taxonomic names.
}
\description{
Prepare enrichment database for matching
}
\details{
This function adds columns to the enrichment database used when matching to taxonomic names. By default this includes:
\itemize{
\item \code{sanitise_name} The sanitised taxonomic name.
\item \code{sanitise_author} The sanitised author of the taxonomic name.
\item \code{require_sanitise} A logical column (TRUE/FALSE) for whether the taxonomic name required sanitising.
= \code{author_parts} Words found in the taxonomic author (removing initials and punctuation) used when performing partial author matching.
\item \code{taxon_length} The length of the taxonomic name (string length) used if typo searching if performed in the matching algorithm.
\item \code{single_entry} A logical column (TRUE/FALSE) for whether the taxonomic name is unique in \code{enrich_database}, used to restrict records when performing either single matching or multiple matching.
\item \code{ID} A column of unique identifiers for each record in \code{enrich_database}. Used for referencing the matches.
}

Moreover the function sorts the enrich database in alphabetical order of the taxonomic names.

These additional column can be switched on/off using the inputs  \code{do_sanitise}, \code{do_taxon_length}, \code{do_single_entry}, \code{do_author_parts} and \code{do_sort}.

Note that if sanitising if performed then sorting and single_entry will be performed with the sanitised taxonoic names and not the inputted names (\code{enrich_database[[enrich_taxon_name_column]]}). Similarly, sanitised authors will be used when creating author parts.
}
\examples{
taxon_names = c('Abies taxifolia', 'ABIES taxifolia',
 'Acalypha gracilens', 'Eupatorium magdalenae', 'Adina racemosa')
taxon_authors = c('Duhamel', 'Poir.', 'A.Gray', 'Stehlé', '(Siebold & Zucc.) Miq.')

enrich_database = data.frame(taxon_names,taxon_authors, value = runif(5))
prepare_enrich_database(enrich_database,
enrich_taxon_name_column = 'taxon_names',
enrich_taxon_authors_column = 'taxon_authors')
}
