% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_functions.R
\name{sanitise_name}
\alias{sanitise_name}
\alias{sanitise_authors}
\alias{sanitise_names_authors}
\alias{sanitise_names_authors_report}
\title{Sanitising functions}
\usage{
sanitise_name(taxon_name)

sanitise_authors(taxon_authors)

sanitise_names_authors(
  taxon_names,
  taxon_authors = NA,
  taxon_names_full = NA,
  console_message = FALSE
)

sanitise_names_authors_report(
  collection,
  taxon_name_column = "TaxonName",
  taxon_name_full_column = NA,
  taxon_author_column = NA,
  console_message = FALSE
)
}
\arguments{
\item{taxon_name}{The taxonomic name of a plant.}

\item{taxon_authors}{Vector of taxonomic authors.}

\item{taxon_names}{Vector of taxonomic names.}

\item{taxon_names_full}{Vector of joined taxonomic name and author.}

\item{console_message}{Flag (TRUE/FALSE) for showing progress bar in the console.}

\item{collection}{A data frame of a collection.}

\item{taxon_name_column}{The name of the column in the \code{collection} corresponding to taxonomic names.}

\item{taxon_name_full_column}{The name of the column in the \code{collection} corresponding to joined taxonomic names and authors.}

\item{taxon_author_column}{The name of the column in the \code{collection} corresponding to the authors of the taxonomic names.}
}
\description{
Functions to clean and standardise taxonomic names and authors.
}
\details{
\code{sanitise_name()} returns the sanitised name of a single taxonomic name.

\code{sanitise_authors()} returns the sanitised name of a  taxonomic authors. Where by characters are coerced to Latin-ASCII, thereby removing diacritics (e.g umlauts).

\code{clean_names_authors()} sanitises multiple taxonomic names with or without the corresponding authors, by applying \code{sanitise_name()} and \code{sanitise_authors()}. As input a vector of taxonomic names is required (\code{taxon_names}), in addition a vector of the authors (\code{taxon_authors}) or joined taxonomic name and authors (\code{taxon_names_full}) can be provided. If neither \code{taxon_authors} or \code{taxon_names_full} are provided the author names are set to \code{''}. A list is returned where:
\itemize{
\item \verb{$taxon_name} a vector of the sanitised taxonomic names,
\item \verb{$author}a vector of the sanitised authors,
\item \verb{$sanitised} is a logical vector of whether the taxon_name was sanitised.
}

\code{clean_names_authors_report()} applies \code{clean_names_authors()} to a collection where the inputs are:
\itemize{
\item a data frame of the collection (\code{collection}),
\item column name for the taxonomic names (\code{taxon_name_column}, required),
\item column name of the authors of the taxonomic names  (\code{taxon_author_column}, optional),
\item column name of the combined taxonomic name and author  (\code{taxon_name_full_column}, optional).
}
}
\examples{
sanitise_name('TRIGONELLA afghanica')
sanitise_name('Halimium X pauanum')
sanitise_name('Aruncus dioicus var acuminatus')
sanitise_authors('Stehlé')

taxon_names = c('TRIGONELLA afghanica', 'Halimium X pauanum',
 'Aruncus dioicus var acuminatus', 'Eupatorium magdalenae')
taxon_authors = c('Vassilcz', 'Font Quer', '(Douglas ex Hook.) H.Hara', 'Stehlé')

sanitise_names_authors(taxon_names, taxon_authors)

collection = data.frame(names = taxon_names, full = paste0(taxon_names, ' ', taxon_authors))
sanitise_names_authors_report(collection, taxon_name_column = 'names',
 taxon_name_full_column = 'full')
}
