% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_type_functions.R
\name{taxon_type}
\alias{taxon_type}
\alias{add_taxon_type}
\title{Extract the taxon type (biological/horticultural characteristics)}
\usage{
taxon_type(taxon_name)

add_taxon_type(
  collection,
  taxon_name_column = "TaxonName",
  POWO_taxon_name_column = NA,
  progress_bar = FALSE
)
}
\arguments{
\item{taxon_name}{The taxonomic name of a plant.}

\item{collection}{A data frame containing a collection.}

\item{taxon_name_column}{The name of the column containing the (original) taxon name.}

\item{POWO_taxon_name_column}{The name of the column containing the POWO (WCVP) Taxon name.}

\item{progress_bar}{Logical flag, if TRUE show progress bar.}
}
\description{
These functions extract the taxon type from a taxon name, in particular whether an item is: an  indeterminate, a species, a subspecies, a variety, a forma, a cultivar or a hybrid.
}
\details{
The aim of these functions to to extract the biological and horticultural characteristics of items via their taxonomic name. For each taxonomic name the functions will return a string containing which taxon types it has. These are:
\itemize{
\item \code{"0.indet"}: The item is indeterminate.
\item \code{"1.species"}: The item is a species.
\item \code{"2.subsp"}: The item is a subspecies.
\item \code{"3.var"}: The item is a variety.
\item \code{"4.f"}: The item is a forma.
\item \code{"5.cultivar"}: The item is a cultivar.
\item \code{"6.hybrid"}: The item is a hybrid.
}

\code{0-4} are biological characteristics and each item can only have one of these characteristics.

\code{5-6} are horticultural characteristics and item can have none, one or both of these.

\code{taxon_type()} returns the taxon type given a single taxonomic name.

\code{add_taxon_type()} adjoins a column called \code{taxon_type} to the original \code{collection} containing the taxon type of each item. If \code{is.na(POWO_taxon_name_column) == TRUE} then the taxon type is extracted purely from the taxon names contained in the column \code{taxon_name_column}. If \code{POWO_taxon_name_column} is a name of a column in \code{collection} then the taxon type is extracted from the column corresponding to it, if there are any missing values then the entries corresponding to \code{taxon_name_column} are used instead.
}
\examples{
taxon_type("Aridaria sp.")
taxon_type("Saxegothaea conspicua")
taxon_type("Rhododendron charitopes subsp. tsangpoense")
taxon_type("Aquilegia flabellata f. alba 'green'")

taxon_names = c("Saintpaulia diplotricha", "Aridaria sp.")
POWO_taxon_name = c("Streptocarpus ionanthus var. diplotrichus", NA)
collection = data.frame(name = taxon_names, POWO = POWO_taxon_name)
add_taxon_type(collection, taxon_name_column = 'name', POWO_taxon_name_column = 'POWO')
}
