% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{typo_list}
\alias{typo_list}
\title{Taxon name typos}
\format{
\subsection{\code{typo_list}}{

A data frame with 8,253 rows and 2 columns:
\describe{
\item{Original}{Original Taxon name with typo}
\item{Fixed}{Fixed Taxon name}
}
}
}
\usage{
typo_list
}
\description{
A data frame of common typos found in the taxon name and their fixes.
}
\keyword{datasets}
