% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_of_items.R
\name{value_of_items}
\alias{value_of_items}
\title{Create value of items in a collection}
\usage{
value_of_items(collection, dependents = NA)
}
\arguments{
\item{collection}{A data frame containing a collection.}

\item{dependents}{A list of scoring systems, where a scoring system is a list of:
\itemize{
\item \verb{$column} the column in the \code{collection} we are generating a score from.
\item \verb{$weight} the weight of this score (multiplier), relative to other columns.
\item \verb{$details} a data frame detailing the score for particular values found in the column. Where the values to match to the column are found in \code{names} and the score is contained in \code{value}.
\item \verb{$type} The method for matching/finding entries in the column. Allowable values are \code{'Exact'}, \code{'Grepl'}, \code{'Number range'}  and \code{'Number range (int)'}.
}}
}
\value{
A list of length two containing:
\itemize{
\item $total_value a vector of length \code{nrow(collection)} containing the value of each item in the collection.
\item $value_breakdown a data frame with \code{nrow(collection)} rows and \code{length(dependents)} columns giving the value of each dependent in corresponding columns.
}
}
\description{
Create value of items in a collection
}
\details{
This function generates a value for each item in a collection given a users scoring system. The scoring system is provided via the input \code{dependents}. Please see the vignette \verb{Adding value to items in a collection} for further documentation.
}
