% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wgsrpd3_level1_simp}
\alias{wgsrpd3_level1_simp}
\title{wgsrpd3_level1_simp}
\format{
\subsection{\code{wgsrpd3_level1_simp}}{

A data frame with 9 rows and 3 columns:
\describe{
\item{code}{The code of the area}
\item{name}{The name of the area}
\item{geometry}{The geometry of the area}
}
}
}
\usage{
wgsrpd3_level1_simp
}
\description{
A data frame containing BRU level 1 areas and their simplified geometries using rmapshaper::ms_simplify.
}
\keyword{datasets}
